+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+
+## [0.1.2]
+
+- Adds support of default config parameter `editor.files` to improve functionality
+- Add first `..` element to allow navigate back in navigation menu
+
 ## [0.1.0]
 
 - Initial release